+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+* Added option `lib` to JuliaCall. Setting this will skip the discovery subprocess.
 * Bug fixes.
 
 ## 0.9.34 (2026-05-18)